DRY up controller Plan Proposal Helper methods
Moving methods that are the same (or mostly the same) to a Concern
module to DRY controller methods. Refs #22715
