+++ components/sponsored_benefits/app/controllers/sponsored_benefits/organizations/plan_design_proposals/benefit_groups_controller.rb
@@ -1,6 +1,8 @@
 module SponsoredBenefits
   module Organizations
     class PlanDesignProposals::BenefitGroupsController < ApplicationController
+      include SponsoredBenefits::Organizations::PlanDesignHelpers
+
       def create
         update_benefit_group_attributes if benefit_group.persisted?
 
@@ -25,28 +27,6 @@ module SponsoredBenefits
       end
 
       private
-        helper_method :plan_design_form, :plan_design_organization, :plan_design_proposal
-
-        def plan_design_proposal
-          @plan_design_proposal ||= SponsoredBenefits::Organizations::PlanDesignProposal.find(params[:plan_design_proposal_id])
-        end
-
-        def plan_design_organization
-          @plan_design_organization ||= plan_design_proposal.plan_design_organization
-        end
-
-        def sponsorship
-          @sponsorship ||= plan_design_proposal.profile.benefit_sponsorships.first
-        end
-
-        def benefit_group
-          @benefit_group ||= sponsorship.benefit_applications.first.benefit_groups.first || sponsorship.benefit_applications.first.benefit_groups.build(benefit_group_params)
-        end
-
-        def plan_design_form
-          SponsoredBenefits::Forms::PlanDesignProposal.new(organization: plan_design_organization, proposal_id: params[:plan_design_proposal_id])
-        end
-
         def update_benefit_group_attributes
           if benefit_group.sole_source?
             benefit_group.composite_tier_contributions.destroy_all
@@ -55,15 +35,6 @@ module SponsoredBenefits
           end
 
           benefit_group.update_attributes(benefit_group_params)
-        end
-
-        def benefit_group_params
-          params.require(:benefit_group).permit(
-                      :reference_plan_id,
-                      :plan_option_kind,
-                      relationship_benefits_attributes: [:relationship, :premium_pct, :offered],
-                      composite_tier_contributions_attributes: [:composite_rating_tier, :employer_contribution_percent, :offered]
-          )
         end
     end
   end
