+++ components/sponsored_benefits/app/controllers/sponsored_benefits/organizations/plan_design_proposals/contributions_controller.rb
@@ -1,6 +1,7 @@
 module SponsoredBenefits
   module Organizations
     class PlanDesignProposals::ContributionsController < ApplicationController
+      include SponsoredBenefits::Organizations::PlanDesignHelpers
 
       def index
         @plan = ::Plan.find(benefit_group.reference_plan_id)
@@ -14,30 +15,6 @@ module SponsoredBenefits
 
         @benefit_group_costs = benefit_group.employee_costs_for_reference_plan
       end
-
-      private
-        helper_method :plan_design_proposal, :benefit_group
-
-        def plan_design_proposal
-          @plan_design_proposal ||= PlanDesignProposal.find(params[:plan_design_proposal_id])
-        end
-
-        def sponsorship
-          @sponsorship ||= plan_design_proposal.profile.benefit_sponsorships.first
-        end
-
-        def benefit_group
-          @benefit_group ||= sponsorship.benefit_applications.first.benefit_groups.build(benefit_group_params)
-        end
-
-        def benefit_group_params
-          params.require(:benefit_group).permit(
-                      :reference_plan_id,
-                      :plan_option_kind,
-                      relationship_benefits_attributes: [:relationship, :premium_pct, :offered],
-                      composite_tier_contributions_attributes: [:composite_rating_tier, :employer_contribution_percent, :offered]
-          )
-        end
     end
   end
 end
