+++ components/sponsored_benefits/app/controllers/sponsored_benefits/organizations/plan_design_proposals/plan_exports_controller.rb
@@ -1,6 +1,7 @@
 module SponsoredBenefits
   module Organizations
     class PlanDesignProposals::PlanExportsController < ApplicationController
+      include SponsoredBenefits::Organizations::PlanDesignHelpers
 
       def new
         @plan_design_organization = plan_design_organization
@@ -21,36 +22,6 @@ module SponsoredBenefits
       end
 
       private
-        helper_method :plan_design_form, :plan_design_organization, :plan_design_proposal, :sponsorship, :visit_types
-
-        def plan_design_proposal
-          @plan_design_proposal ||= SponsoredBenefits::Organizations::PlanDesignProposal.find(params[:plan_design_proposal_id])
-        end
-
-        def plan_design_organization
-          @plan_design_organization ||= plan_design_proposal.plan_design_organization
-        end
-
-        def sponsorship
-          @sponsorship ||= plan_design_proposal.profile.benefit_sponsorships.first
-        end
-
-        def benefit_group
-          @benefit_group ||= sponsorship.benefit_applications.first.benefit_groups.first || sponsorship.benefit_applications.first.benefit_groups.build(benefit_group_params)
-        end
-
-        def plan_array(plan)
-           ::Plan.where(:_id => { '$in': [plan.id] } ).map(&:hios_id)
-        end
-
-        def plan_design_form
-          SponsoredBenefits::Forms::PlanDesignProposal.new(organization: plan_design_organization, proposal_id: params[:plan_design_proposal_id])
-        end
-
-        def visit_types
-          @visit_types ||= ::Products::Qhp::VISIT_TYPES
-        end
-
         def sbc_included
           params[:sbc_included] == 'true'
         end
